+++ jupiter-registry/jupiter-registry-api/src/main/java/org/jupiter/registry/AbstractRegistryService.java
@@ -20,14 +20,13 @@ import org.jupiter.common.concurrent.NamedThreadFactory;
 import org.jupiter.common.concurrent.collection.ConcurrentSet;
 import org.jupiter.common.util.Lists;
 import org.jupiter.common.util.Maps;
-import org.jupiter.common.util.Pair;
 import org.jupiter.common.util.internal.logging.InternalLogger;
 import org.jupiter.common.util.internal.logging.InternalLoggerFactory;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.locks.Lock;
@@ -50,8 +49,8 @@ public abstract class AbstractRegistryService implements RegistryService {
             Executors.newSingleThreadExecutor(new NamedThreadFactory("registry.executor"));
     private final AtomicBoolean shutdown = new AtomicBoolean(false);
 
-    private final Map<RegisterMeta.ServiceMeta, Pair<Long, List<RegisterMeta>>> registries = Maps.newHashMap();
-    private final ReentrantReadWriteLock registriesLock = new ReentrantReadWriteLock();
+    private final ConcurrentMap<RegisterMeta.ServiceMeta, RegisterValue> registries =
+            Maps.newConcurrentMap();
 
     private final ConcurrentMap<RegisterMeta.ServiceMeta, CopyOnWriteArrayList<NotifyListener>> subscribeListeners =
             Maps.newConcurrentMap();
@@ -116,20 +115,19 @@ public abstract class AbstractRegistryService implements RegistryService {
 
     @Override
     public Collection<RegisterMeta> lookup(RegisterMeta.ServiceMeta serviceMeta) {
-        Pair<Long, List<RegisterMeta>> data;
+        RegisterValue value = registries.get(serviceMeta);
 
-        final Lock readLock = registriesLock.readLock();
+        if (value == null) {
+            return Collections.emptyList();
+        }
+
+        final Lock readLock = value.lock.readLock();
         readLock.lock();
         try {
-            data = registries.get(serviceMeta);
+            return Lists.newArrayList(value.metaSet);
         } finally {
             readLock.unlock();
         }
-
-        if (data != null) {
-            return Lists.newArrayList(data.getSecond());
-        }
-        return Collections.emptyList();
     }
 
     @Override
@@ -187,36 +185,32 @@ public abstract class AbstractRegistryService implements RegistryService {
             return;
         }
 
+        RegisterValue value = registries.get(serviceMeta);
+
+        if (value == null) {
+            RegisterValue newValue = new RegisterValue();
+            value = registries.putIfAbsent(serviceMeta, newValue);
+            if (value == null) {
+                value = newValue;
+            }
+        }
+
         boolean notifyNeeded = false;
 
-        final Lock writeLock = registriesLock.writeLock();
+        final Lock writeLock = value.lock.writeLock();
         writeLock.lock();
         try {
-            Pair<Long, List<RegisterMeta>> data = registries.get(serviceMeta);
-            if (data == null) {
+            if (version > value.version) {
                 if (event == NotifyListener.NotifyEvent.CHILD_REMOVED) {
-                    return;
-                }
-                List<RegisterMeta> metaList = Lists.newArrayList(array);
-                data = Pair.of(version, metaList);
-                notifyNeeded = true;
-            } else {
-                long oldVersion = data.getFirst();
-                List<RegisterMeta> metaList = data.getSecond();
-                if (oldVersion < version || (version < 0 && oldVersion > 0 /* version 溢出 */)) {
-                    if (event == NotifyListener.NotifyEvent.CHILD_REMOVED) {
-                        for (RegisterMeta m : array) {
-                            metaList.remove(m);
-                        }
-                    } else if (event == NotifyListener.NotifyEvent.CHILD_ADDED) {
-                        Collections.addAll(metaList, array);
+                    for (RegisterMeta m : array) {
+                        value.metaSet.remove(m);
                     }
-                    data = Pair.of(version, metaList);
-                    notifyNeeded = true;
+                } else if (event == NotifyListener.NotifyEvent.CHILD_ADDED) {
+                    Collections.addAll(value.metaSet, array);
                 }
+                value.version = version;
+                notifyNeeded = true;
             }
-
-            registries.put(serviceMeta, data);
         } finally {
             writeLock.unlock();
         }
@@ -238,4 +232,10 @@ public abstract class AbstractRegistryService implements RegistryService {
     protected abstract void doRegister(RegisterMeta meta);
 
     protected abstract void doUnregister(RegisterMeta meta);
+
+    private static class RegisterValue {
+        private long version = Long.MIN_VALUE;
+        private final Set<RegisterMeta> metaSet = new HashSet<>();
+        private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
+    }
 }